Remove title from <iframe /> container
On Firefox, a tooltip appears when hovering clickable elements inside the iframe with the "title" attribute of the iframe. Hovering over the an action button "Submit", for example, would show "Embedded Portal Content", which was distracting. Removing the title should solve the issue.

On Chrome, this tooltip was visible only when hovering the edges of the iframe, not the content itself.